test-requirements.txt
coverage==3.7.1
Flask-Testing==0.4.2
nose==1.3.4
